Move `maybeDecompressRuntimeBlob` in a web worker
