+++ packages/ui/src/common/utils/crypto/worker/utils.ts
@@ -1,24 +1,26 @@
-import { merkleRootFromBinary, hashFile } from '..'
+import { merkleRootFromBinary, hashFile, maybeDecompressRuntimeBlob } from '..'
 
-export type WorkerRequestType = 'HASH_FILE' | 'MERKLE_ROOT'
+export type WorkerRequestType = 'HASH_FILE' | 'MERKLE_ROOT' | 'DECOMPRESS_RUNTIME'
 export type WorkerRequest = {
   type: WorkerRequestType
   id: string
-  file: Blob
+  file: Blob | ArrayBuffer
 }
 
 export type WorkerResponse = {
   type: WorkerRequestType
   id: string
-  value: string
+  value: any
   error?: boolean
 }
 
-export const compute = async (type: WorkerRequestType, file: Blob): Promise<string> => {
+export async function compute(type: WorkerRequestType, file: Blob | ArrayBuffer): Promise<any> {
   switch (type) {
     case 'HASH_FILE':
-      return await hashFile(file)
+      return await hashFile(file as Blob)
     case 'MERKLE_ROOT':
-      return await merkleRootFromBinary(file)
+      return await merkleRootFromBinary(file as Blob)
+    case 'DECOMPRESS_RUNTIME':
+      return await maybeDecompressRuntimeBlob(file as ArrayBuffer)
   }
 }
